JupyterHub OIDC authentication method
This commit provides a Python class called `RucioAuthenticator` extended from JupyterHubs `GenericOAuthenticator` to fetch the access token from the `auth_state` of the user, request a new refresh token, and place it into the spawners environment for the
extension to access. It also provides the rest of the helm config for the setup. Additional documentation in the readme.

ref #20